requirements.txt
onnxruntime
opencv-python < 4.8.0
numpy
pillow
pytest